+++ polars/polars-lazy/src/physical_plan/expressions/default.rs
@@ -9,34 +9,6 @@ use rayon::prelude::*;
 use std::borrow::Cow;
 use std::sync::Arc;
 
-pub struct IsNullExpr {
-    physical_expr: Arc<dyn PhysicalExpr>,
-    expr: Expr,
-}
-
-impl IsNullExpr {
-    pub fn new(physical_expr: Arc<dyn PhysicalExpr>, expr: Expr) -> Self {
-        Self {
-            physical_expr,
-            expr,
-        }
-    }
-}
-
-impl PhysicalExpr for IsNullExpr {
-    fn as_expression(&self) -> &Expr {
-        &self.expr
-    }
-
-    fn evaluate(&self, df: &DataFrame, state: &ExecutionState) -> Result<Series> {
-        let series = self.physical_expr.evaluate(df, state)?;
-        Ok(series.is_null().into_series())
-    }
-    fn to_field(&self, _input_schema: &Schema) -> Result<Field> {
-        Ok(Field::new("is_null", DataType::Boolean))
-    }
-}
-
 pub struct IsNotNullExpr {
     physical_expr: Arc<dyn PhysicalExpr>,
     expr: Expr,
